Create olcPGEX_GUI.hxx extension
This is what I have in mind when I said that I will create a GUI library extension of olcPixelGameEngine. Although it might take a while to finish since I have been focusing on school recently and I have other projects to do. So if any of you are interested, you can contribute ideas or bits of code to this project.

-Clyde