Corrected Linux numpad key mapping
I guess there was a problem with the numpad requiring "XLookupString" and found a simple fix while working on a project. The "right" keycodes aliases does not represent the digits but the alt+key functions (5 corresponds to as spooky ghost "begin" key). The previous solution was also leading to conflicting keycodes with non QWERTY keyboards, uppercase letters registering as qwerty but lowercase as AZERTY, two different keys updated at the same time by only pressing one, etc...

https://cgit.freedesktop.org/xorg/proto/x11proto/tree/keysymdef.h

I implemented my mapkey with a switch/case inside a function since I was programming it in C (no "map" structure), it may be a cleaner way to do it afterall.