Set Engine Thread name on Windows
Windows debuggers can be informed with the name of thread using a
special exception type. This adds a new olc::Platform api to allow
setting a thread name, implmements it for windows, and calls it from the
enigne thread.

This is just a convenience for debugging on windows. It will cause the
VS Debugger to show the thread as "OLC PGE Engine Thread" as opposed to
the default "ucrtbased.dll thread"